services/content_moderation/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
torch>=2.0.0
transformers>=4.40.0
Pillow>=9.0.0
open_clip_torch>=2.23.0
python-multipart==0.0.20 